Restructure project to allow more commands
