+++ lib/corneal.rb
@@ -1,0 +1,9 @@
+require "extensions/string"
+require "corneal/generators/app/app_generator"
+# require "corneal/generators/model/model_generator"
+require "corneal/version"
+require "corneal/cli"
+
+module Corneal
+
+end
