+++ lib/corneal/generators/app/templates/Gemfile
