+++ lib/corneal/generators/app/templates/README.md
