+++ lib/corneal/generators/app/templates/Rakefile
