+++ lib/corneal/generators/app/templates/app/controllers/application_controller.rb
