+++ lib/corneal/generators/app/templates/app/views/layout.erb.tt
