+++ lib/corneal/generators/app/templates/app/views/welcome.erb
