+++ lib/corneal/generators/app/templates/config/db.yml
