+++ lib/corneal/generators/app/templates/config/environment.rb
