+++ lib/corneal/generators/app/templates/config/initializers/database.rb
