+++ lib/corneal/generators/app/templates/config/initializers/redis.rb
