+++ lib/corneal/generators/app/templates/config/redis.yml
