+++ lib/corneal/generators/app/templates/public/favicon.ico
