+++ lib/corneal/generators/app/templates/public/images/corneal-small.png
