+++ lib/corneal/generators/app/templates/public/stylesheets/main.css
