+++ lib/corneal/generators/app/templates/spec/application_controller_spec.rb
