+++ lib/corneal/generators/app/templates/spec/spec_helper.rb
